Extract path builder on its own
